rename AnimalForm and VegetationForm to AnimalFormData and VegetationFormData
