+++ apps/thundermoon_web/lib/thundermoon_web/live/lotka_volterra_live/new_predator.ex
@@ -2,7 +2,7 @@ defmodule ThundermoonWeb.LotkaVolterraLive.NewPredator do
   use ThundermoonWeb.Component.EntityForm,
     params_name: "predator",
     model: LotkaVolterra.Predator,
-    form_data: Thundermoon.LotkaVolterra.AnimalForm
+    form_data: Thundermoon.LotkaVolterra.AnimalFormData
 
   alias LotkaVolterra.Predator
